+++ tests/CloudinaryImageHost/DestroyAllTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace PZL\SiteImage\Tests\CloudinaryHost;
+namespace PZL\SiteImage\Tests\CloudinaryImageHost;
 
 
 use PZL\SiteImage\Tests\TestCase;
 
-class GetTransformationsTest extends TestCase
+class DestroyAllTest extends TestCase
 {
 
     public function testToBeImplemented()