+++ tests/LocalImageHost/UploadTest.php
@@ -3,10 +3,10 @@
  * Copyright (c) 2022 Perfect Zero Labs.
  */
 
-namespace PZL\SiteImage\Tests\LocalHost;
+namespace PZL\SiteImage\Tests\LocalImageHost;
 
 use Illuminate\Filesystem\Filesystem;
-use PZL\SiteImage\Host\LocalHost;
+use PZL\SiteImage\Host\LocalImageHost;
 use PZL\SiteImage\Tests\TestCase;
 
 /**
@@ -15,7 +15,7 @@ use PZL\SiteImage\Tests\TestCase;
 class UploadTest extends TestCase
 {
     /**
-     * @var LocalHost()
+     * @var LocalImageHost()
      */
     private $provider;
 
@@ -33,7 +33,7 @@ class UploadTest extends TestCase
     {
         parent::setUp();
 
-        $this->provider = new LocalHost();
+        $this->provider = new LocalImageHost();
 
         $fs = new Filesystem();
         $fs->cleanDirectory($this->provider->getFolder());